add percentage to reports

+++ Template.Application/Reports/Dtos/ComplaintGovermentalEntitiesReportDto.cs
@@ -5,6 +5,6 @@
         public int GovermentalEntityId { get; set; }
         public string GovermentalEntityName { get; set; } = string.Empty;
         public int ComplaintCount { get; set; }
-        public int PercentageOfTotalComplaints { get; set; }
+        public decimal PercentageOfTotalComplaints { get; set; }
     }
 }
