+++ Template.Application/Reports/Dtos/ComplaintTimeBasedReportDto.cs
@@ -5,6 +5,6 @@
         public int Year { get; set; }
         public int Month { get; set; }
         public int ComplaintCount { get; set; }
-        public int PercentageOfTotalComplaints { get; set; }
+        public decimal PercentageOfTotalComplaints { get; set; }
     }
 }
