+++ Template.Application/Reports/Dtos/ComplaintsStatusReportCountDto.cs
@@ -6,6 +6,6 @@ public class ComplaintsStatusReportCountDto
 {
     public ComplaintStatus Status { get; set; }
     public int ComplaintCount { get; set; }
-    public int PercentageOfTotalComplaints { get; set; }
+    public decimal PercentageOfTotalComplaints { get; set; }
 
 }
